+++ Runtime/Config/VFXLibrary.cs
@@ -1,18 +1,24 @@
 using System;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.Serialization;
 
 namespace WTFGames.Hephaestus.VFX {
     [CreateAssetMenu(fileName = "VFXLibrary", menuName = "HephaestusMobile/Core/VFX/VFXLibrary")]
     public class VFXLibrary : ScriptableObject {
-        
-        public VFXLibraryConstants widgetsLibraryConstants;
-        
+
+        [FormerlySerializedAs("widgetsLibraryConstants")]
+        public VFXLibraryConstants vfxLibraryConstants;
+
         [HideInInspector]
         public List<VFXNamePair> vfxList = new List<VFXNamePair>();
-        
+
+        /// <summary>
+        /// Returns the prefab mapped to the given VFX type, or null when there is no mapping.
+        /// </summary>
         public GameObject GetPrefabByType(Enum vfxType) {
-            return vfxList.Find(w => w.vfxType == Convert.ToInt32(vfxType)).vfxPrefab;
+            var vfxTypeId = Convert.ToInt32(vfxType);
+            return vfxList.Find(w => w.vfxType == vfxTypeId)?.vfxPrefab;
         }
     }
 }