+++ Runtime/Manager/IVFXManager.cs
@@ -3,19 +3,13 @@ using UnityEngine;
 
 namespace WTFGames.Hephaestus.VFX {
     public interface IVFXManager {
-        
+
         /// <summary>
-        /// Initialize the VFX manager.
-        /// </summary>
-        /// <param name="vfxManagerConfig">VFXManagerConfig file.</param>
-        void Initialize();
-        
-        /// <summary>
-        /// Insttantiate and play VFX entity.
+        /// Instantiate and play VFX entity.
         /// </summary>
         /// <param name="vfxName">VFX name.</param>
-        /// <param name="position">Desired position.</param>
+        /// <param name="position">Desired position in world space.</param>
         /// <param name="parent">Parent object.</param>
-        void PlayVFX(Enum vfxName, Vector3 position, Transform parent = null);
+        void PlayVFX(Enum vfxName, Vector3 position = default, Transform parent = null);
     }
 }